Develop (#16)
* Configuring code climate for Java code analysis

* Testing Checkstyle

* Playing with coverity settings

* Playing with coverity settings

* Playing with coverity settings

* Playing with coverity settings

* Playing with coverity settings

* Playing with coverity settings

* Playing with Code climate PMD

* Playing with Code climate - testing Checkstyle

* Playing with Code climate - testing Checkstyle

* Playing with Code climate - testing Checkstyle

* Playing with Code climate - testing Checkstyle

* Playing with Code climate - testing Checkstyle

* Remove codeclimate

* Removed duplicate code resulting from bad merge

* Added missing test annotations

* Added missing test annotations

* Feature/unit tests (#10)

* Merge develop into master

* Added missing annotations

* Added more missing annotations

* Migrated test suite from JUnit to TestNG and parallelized tests

* Expanded ArrayMath tests

* Update README.md

* Added more Arraymath Tests

* Create CONTRIBUTING.md

* Added more Arraymath Tests

* Added more Arraymath Tests

* Merge branch 'master' of https://github.com/MinesJTK/jtk into feature/Unit_Tests

# Conflicts:
#	core/src/test/java/edu/mines/jtk/util/ArrayMathTest.java

* Merge branch 'master' of https://github.com/MinesJTK/jtk into feature/Unit_Tests

# Conflicts:
#	core/src/test/java/edu/mines/jtk/util/ArrayMathTest.java

* Merge branch 'master' of https://github.com/MinesJTK/jtk into feature/Unit_Tests

# Conflicts:
#	core/src/test/java/edu/mines/jtk/util/ArrayMathTest.java

* Merge branch 'master' of https://github.com/MinesJTK/jtk into feature/Unit_Tests

# Conflicts:
#	core/src/test/java/edu/mines/jtk/util/ArrayMathTest.java

* Merge branch 'master' of https://github.com/MinesJTK/jtk into feature/Unit_Tests

# Conflicts:
#	core/src/test/java/edu/mines/jtk/util/ArrayMathTest.java

* Merge branch 'master' of https://github.com/MinesJTK/jtk into feature/Unit_Tests

# Conflicts:
#	core/src/test/java/edu/mines/jtk/util/ArrayMathTest.java

* Mjavaerge branch 'master' of https://github.com/MinesJTK/jtk into feature/Unit_Tests

# Conflicts:
#	core/src/test/java/edu/mines/jtk/util/ArrayMathTest.

* Mjavaerge branch 'master' of https://github.com/MinesJTK/jtk into feature/Unit_Tests

# Conflicts:
#	core/src/test/java/edu/mines/jtk/util/ArrayMathTest.

* Mjavaerge branch 'master' of https://github.com/MinesJTK/jtk into feature/Unit_Tests

# Conflicts:
#	core/src/test/java/edu/mines/jtk/util/ArrayMathTest.

* Adjusted testing variables scope to private

* Adjust parallel testing to class-level from method-level

* Adjust parallel testing to class-level from method-level

* Adjust parallel testing to class-level from method-level

* Adjust parallel testing to class-level from method-level

* Update travis file and refactored ArrayMathTests

* Quick fix for byte array math tests
