+++ .travis.yml
@@ -21,14 +21,12 @@ cache:
     - $HOME/.gradle/wrapper/
 
 before_script:
-  - "chmod +x gradlew"
   - "export DISPLAY=:99.0"
   - "sh -e /etc/init.d/xvfb start"
   - sleep 3
   - "/sbin/start-stop-daemon --start --quiet --pidfile /tmp/custom_xvfb_99.pid --make-pidfile --background --exec /usr/bin/Xvfb -- :99 -ac -screen 0 1280x1024x16 +extension RANDR"
 
-script:
-  - "./gradlew build -x :core:signArchives"
+script: gradle build
 
 after_success:
   - bash <(curl -s https://codecov.io/bash)
