+++ core/build.gradle
@@ -72,49 +72,3 @@ artifacts {
   archives javadocJar
   archives sourcesJar
 }
-
-signing {
-  sign configurations.archives
-}
-
-if (project.hasProperty('sonatypeUsername')) {
-  uploadArchives {
-    repositories {
-      mavenDeployer {
-        beforeDeployment { MavenDeployment deployment -> signing.signPom(deployment) }
-
-        repository(url: "https://oss.sonatype.org/service/local/staging/deploy/maven2/") {
-          authentication(userName: sonatypeUsername, password: sonatypePassword)
-        }
-
-        pom.project {
-          name 'Mines Java Toolkit'
-          packaging 'jar'
-          description 'Java packages for science and engineering'
-          url 'https://github.com/MinesJTK/jtk/'
-
-          scm {
-            url 'scm:git@github.com/MinesJTK/jtk.git'
-            connection 'scm:git@github.com/MinesJTK/jtk.git'
-            developerConnection 'scm:git@github.com/MinesJTK/jtk.git'
-          }
-
-          licenses {
-            license {
-              name 'The Apache Software License, Version 2.0'
-              url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-              distribution 'repo'
-            }
-          }
-
-          developers {
-            developer {
-              id 'dhale'
-              name 'Dave Hale'
-            }
-          }
-        }
-      }
-    }
-  }
-}
