+++ core/src/test/java/edu/mines/jtk/util/ArrayMathFloatTest.java
@@ -14,29 +14,39 @@
  ****************************************************************************/
 package edu.mines.jtk.util;
 
+import org.testng.AssertJUnit;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
-
 import java.util.Random;
 
 import static edu.mines.jtk.util.ArrayMath.*;
-import static org.testng.Assert.*;
+import static edu.mines.jtk.util.ArrayMath.transpose;
+
+import static org.testng.Assert.assertFalse;
+import static org.testng.Assert.assertEquals;
+import static org.testng.Assert.assertTrue;
+
+import static edu.mines.jtk.util.ArrayMathTest.*;
 
 /**
- * Tests double array operations in {@link edu.mines.jtk.util.ArrayMath}.
+ * Tests float array operations in {@link edu.mines.jtk.util.ArrayMath}.
  * @author Chris Engelsma
  * @version 2017.05.31
  */
-public class ArrayMathTestDouble extends ArrayMathTest {
+public class ArrayMathFloatTest {
+
+  private static final int n3 = 8;
+  private static final int n2 = 6;
+  private static final int n1 = 4;
 
   @BeforeMethod
   public void setUp() {
-    a1 = rampdouble(0d,1d,n1);
-    a2 = rampdouble(0d,1d,10d,n1,n2);
-    a3 = rampdouble(0d,1d,10d,100d,n1,n2,n3);
-    b1 = zerodouble(n1);
-    b2 = zerodouble(n1,n2);
-    b3 = zerodouble(n1,n2,n3);
+    a1 = rampfloat(0,1,n1);
+    a2 = rampfloat(0,1,10,n1,n2);
+    a3 = rampfloat(0,1,10,100,n1,n2,n3);
+    b1 = zerofloat(n1);
+    b2 = zerofloat(n1,n2);
+    b3 = zerofloat(n1,n2,n3);
   }
 
   @Test
@@ -47,18 +57,18 @@ public class ArrayMathTestDouble extends ArrayMathTest {
     fill(1,a3);
     assertFalse(equal(a3,b3));
 
-    assertTrue(equal(1.0,a3,b3));
-    assertFalse(equal(0.99,a3,b3));
+    assertTrue(equal(1.0f,a3,b3));
+    assertFalse(equal(0.99f,a3,b3));
   }
 
   @Test
-  public void testRampDoubles() {
+  public void testRampFloats() {
     assertArraySize(a1,n1);
     assertArraySize(a2,n1,n2);
     assertArraySize(a3,n1,n2,n3);
 
     for (int i1=0; i1<n1; ++i1) {
-      double expected = (double)i1;
+      float expected = (float)i1;
       assertEquals(expected,a3[0][0][i1]);
       assertEquals(expected,a2[0][i1]);
       assertEquals(expected,a1[i1]);
@@ -66,14 +76,14 @@ public class ArrayMathTestDouble extends ArrayMathTest {
   }
 
   @Test
-  public void testFillDoubles() {
-    double val = 1.0d;
+  public void testFillFloats() {
+    float val = 1.0f;
 
-    assertEqual(zerodouble(n1,n2,n3),filldouble(0.0,n1,n2,n3));
+    assertEqual(zerofloat(n1,n2,n3),fillfloat(0.0f,n1,n2,n3));
 
-    a1 = filldouble(val,n1);
-    a2 = filldouble(val,n1,n2);
-    a3 = filldouble(val,n1,n2,n3);
+    a1 = fillfloat(val,n1);
+    a2 = fillfloat(val,n1,n2);
+    a3 = fillfloat(val,n1,n2,n3);
 
     assertArraySize(a1,n1);
     assertArraySize(a2,n1,n2);
@@ -103,9 +113,9 @@ public class ArrayMathTestDouble extends ArrayMathTest {
     assertArraySize(a2,n1,n2);
     assertArraySize(a3,n1,n2,n3);
 
-    b1 = randdouble(n1);
-    b2 = randdouble(n1,n2);
-    b3 = randdouble(n1,n2,n3);
+    b1 = randfloat(n1);
+    b2 = randfloat(n1,n2);
+    b3 = randfloat(n1,n2,n3);
 
     assertArraySize(b1,n1);
     assertArraySize(b2,n1,n2);
@@ -115,7 +125,6 @@ public class ArrayMathTestDouble extends ArrayMathTest {
     assertFalse(equal(a2,b2));
     assertFalse(equal(a3,b3));
   }
-
   @Test
   public void testCopySimple() {
     b1 = copy(a1);
@@ -132,50 +141,50 @@ public class ArrayMathTestDouble extends ArrayMathTest {
     assertEqual(b2,a2);
     assertEqual(b3,a3);
 
-    b1 = copy(n1 - 1,a1);
-    b2 = copy(n1 - 1,n2 - 1,a2);
-    b3 = copy(n1 - 1,n2 - 1,n3 - 1,a3);
-    assertEqual(b1,rampdouble(0,1,n1 - 1));
-    assertEqual(b2,rampdouble(0,1,10,n1 - 1,n2 - 1));
-    assertEqual(b3,rampdouble(0,1,10,100,n1 - 1,n2 - 1,n3 - 1));
+    b1 = copy(n1-1,a1);
+    b2 = copy(n1-1,n2-1,a2);
+    b3 = copy(n1-1,n2-1,n3-1,a3);
+    assertEqual(b1,rampfloat(0f,1f,n1-1));
+    assertEqual(b2,rampfloat(0f,1f,10f,n1-1,n2-1));
+    assertEqual(b3,rampfloat(0f,1f,10f,100f,n1-1,n2-1,n3-1));
 
-    copy(n1 - 1,a1,b1);
-    copy(n1 - 1,n2 - 1,a2,b2);
-    copy(n1 - 1,n2 - 1,n3 - 1,a3,b3);
-    assertEqual(b1,rampdouble(0,1,n1 - 1));
-    assertEqual(b2,rampdouble(0,1,10,n1 - 1,n2 - 1));
-    assertEqual(b3,rampdouble(0,1,10,100,n1 - 1,n2 - 1,n3 - 1));
+    copy(n1-1,a1,b1);
+    copy(n1-1,n2-1,a2,b2);
+    copy(n1-1,n2-1,n3-1,a3,b3);
+    assertEqual(b1,rampfloat(0f,1f,n1-1));
+    assertEqual(b2,rampfloat(0f,1f,10f,n1-1,n2-1));
+    assertEqual(b3,rampfloat(0f,1f,10f,100f,n1-1,n2-1,n3-1));
 
-    b1 = copy(n1 - 1,1,a1);
-    b2 = copy(n1 - 2,n2 - 1,2,1,a2);
-    b3 = copy(n1 - 3,n2 - 2,n3 - 1,3,2,1,a3);
-    assertEqual(b1,rampdouble(1,1,n1 - 1));
-    assertEqual(b2,rampdouble(12,1,10,n1 - 1,n2 - 1));
-    assertEqual(b3,rampdouble(123,1,10,100,n1 - 1,n2 - 1,n3 - 1));
+    b1 = copy(n1-1,1,a1);
+    b2 = copy(n1-2,n2-1,2,1,a2);
+    b3 = copy(n1-3,n2-2,n3-1,3,2,1,a3);
+    assertEqual(b1,rampfloat(1f,1f,n1-1));
+    assertEqual(b2,rampfloat(12f,1f,10f,n1-1,n2-1));
+    assertEqual(b3,rampfloat(123f,1f,10f,100f,n1-1,n2-1,n3-1));
 
-    copy(n1 - 1,1,a1,0,b1);
-    copy(n1 - 2,n2 - 1,2,1,a2,0,0,b2);
-    copy(n1 - 3,n2 - 2,n3 - 1,3,2,1,a3,0,0,0,b3);
-    assertEqual(b1,rampdouble(1,1,n1 - 1));
-    assertEqual(b2,rampdouble(12,1,10,n1 - 1,n2 - 1));
-    assertEqual(b3,rampdouble(123,1,10,100,n1 - 1,n2 - 1,n3 - 1));
+    copy(n1-1,1,a1,0,b1);
+    copy(n1-2,n2-1,2,1,a2,0,0,b2);
+    copy(n1-3,n2-2,n3-1,3,2,1,a3,0,0,0,b3);
+    assertEqual(b1,rampfloat(1f,1f,n1-1));
+    assertEqual(b2,rampfloat(12f,1f,10f,n1-1,n2-1));
+    assertEqual(b3,rampfloat(123f,1f,10f,100f,n1-1,n2-1,n3-1));
 
-    b1 = copy(n1 / 2,0,2,a1);
-    b2 = copy(n1 / 2,n2 / 2,0,0,2,2,a2);
-    b3 = copy(n1 / 2,n2 / 2,n3 / 2,0,0,0,2,2,2,a3);
-    assertEqual(b1,rampdouble(0,2,n1 / 2));
-    assertEqual(b2,rampdouble(0,2,20,n1 / 2,n2 / 2));
-    assertEqual(b3,rampdouble(0,2,20,200,n1 / 2,n2 / 2,n3 / 2));
+    b1 = copy(n1/2,0,2,a1);
+    b2 = copy(n1/2,n2/2,0,0,2,2,a2);
+    b3 = copy(n1/2,n2/2,n3/2,0,0,0,2,2,2,a3);
+    assertEqual(b1,rampfloat(0f,2f,n1/2));
+    assertEqual(b2,rampfloat(0f,2f,20f,n1/2,n2/2));
+    assertEqual(b3,rampfloat(0f,2f,20f,200f,n1/2,n2/2,n3/2));
 
     b1 = copy(a1);
     b2 = copy(a2);
     b3 = copy(a3);
-    copy(n1 - 1,1,a1,1,b1);
-    copy(n1 - 2,n2 - 1,2,1,a2,2,1,b2);
-    copy(n1 - 3,n2 - 2,n3 - 1,3,2,1,a3,3,2,1,b3);
-    assertEqual(b1,rampdouble(0,1,n1));
-    assertEqual(b2,rampdouble(0,1,10,n1,n2));
-    assertEqual(b3,rampdouble(0,1,10,100,n1,n2,n3));
+    copy(n1-1,1,a1,1,b1);
+    copy(n1-2,n2-1,2,1,a2,2,1,b2);
+    copy(n1-3,n2-2,n3-1,3,2,1,a3,3,2,1,b3);
+    assertEqual(b1,rampfloat(0f,1f,n1));
+    assertEqual(b2,rampfloat(0f,1f,10f,n1,n2));
+    assertEqual(b3,rampfloat(0f,1f,10f,100f,n1,n2,n3));
   }
 
   @Test
@@ -207,9 +216,9 @@ public class ArrayMathTestDouble extends ArrayMathTest {
     assertFalse(distinct(a2,b2));
     assertFalse(distinct(a1,b1));
 
-    b3 = new double[n3][n2][n1];
-    b2 = new double[n2][n1];
-    b1 = new double[n1];
+    b3 = new float[n3][n2][n1];
+    b2 = new float[n2][n1];
+    b1 = new float[n1];
     assertTrue(distinct(a3,b3));
     assertTrue(distinct(a2,b2));
     assertTrue(distinct(a1,b1));
@@ -222,16 +231,16 @@ public class ArrayMathTestDouble extends ArrayMathTest {
 
   @Test
   public void testAddition() {
-    b1 = rampdouble(1,1,n1);                assertEqual(b1,add(1,a1));
+    b1 = rampfloat(1,1,n1);                 assertEqual(b1,add(1,a1));
                                             assertEqual(b1,add(a1,1));
-    b2 = rampdouble(1,1,10,n1,n2);          assertEqual(b2,add(1,a2));
+    b2 = rampfloat(1,1,10,n1,n2);           assertEqual(b2,add(1,a2));
                                             assertEqual(b2,add(a2,1));
-    b3 = rampdouble(1,1,10,100,n1,n2,n3);   assertEqual(b3,add(1,a3));
+    b3 = rampfloat(1,1,10,100,n1,n2,n3);    assertEqual(b3,add(1,a3));
                                             assertEqual(b3,add(a3,1));
 
-    b1 = rampdouble(0,2,n1);                assertEqual(b1,add(a1,a1));
-    b2 = rampdouble(0,2,20,n1,n2);          assertEqual(b2,add(a2,a2));
-    b3 = rampdouble(0,2,20,200,n1,n2,n3);   assertEqual(b3,add(a3,a3));
+    b1 = rampfloat(0,2,n1);                 assertEqual(b1,add(a1,a1));
+    b2 = rampfloat(0,2,20,n1,n2);           assertEqual(b2,add(a2,a2));
+    b3 = rampfloat(0,2,20,200,n1,n2,n3);    assertEqual(b3,add(a3,a3));
 
     add(a1,a1,a1);                          assertEquals(b1,a1);
     add(a2,a2,a2);                          assertEquals(b2,b2);
@@ -240,13 +249,13 @@ public class ArrayMathTestDouble extends ArrayMathTest {
 
   @Test
   public void testSubtraction() {
-    b1 = rampdouble(-1,1,n1);               assertEqual(b1,sub(a1,1));
-    b2 = rampdouble(-1,1,10,n1,n2);         assertEqual(b2,sub(a2,1));
-    b3 = rampdouble(-1,1,10,100,n1,n2,n3);  assertEqual(b3,sub(a3,1));
+    b1 = rampfloat(-1,1,n1);                assertEqual(b1,sub(a1,1));
+    b2 = rampfloat(-1,1,10,n1,n2);          assertEqual(b2,sub(a2,1));
+    b3 = rampfloat(-1,1,10,100,n1,n2,n3);   assertEqual(b3,sub(a3,1));
 
-    b1 = rampdouble(1,-1,n1);               assertEqual(b1,sub(1,a1));
-    b2 = rampdouble(1,-1,-10,n1,n2);        assertEqual(b2,sub(1,a2));
-    b3 = rampdouble(1,-1,-10,-100,n1,n2,n3);assertEqual(b3,sub(1,a3));
+    b1 = rampfloat(1,-1,n1);                assertEqual(b1,sub(1,a1));
+    b2 = rampfloat(1,-1,-10,n1,n2);         assertEqual(b2,sub(1,a2));
+    b3 = rampfloat(1,-1,-10,-100,n1,n2,n3); assertEqual(b3,sub(1,a3));
 
     zero(b1);                               assertEqual(b1,sub(a1,a1));
     zero(b2);                               assertEqual(b2,sub(a2,a2));
@@ -259,22 +268,22 @@ public class ArrayMathTestDouble extends ArrayMathTest {
 
   @Test
   public void testMultiplication() {
-    b1 = rampdouble(0,2,n1);                assertEqual(b1,mul(2,a1));
+    b1 = rampfloat(0,2,n1);                 assertEqual(b1,mul(2,a1));
                                             assertEqual(b1,mul(a1,2));
-    b2 = rampdouble(0,2,20,n1,n2);          assertEqual(b2,mul(2,a2));
+    b2 = rampfloat(0,2,20,n1,n2);           assertEqual(b2,mul(2,a2));
                                             assertEqual(b2,mul(a2,2));
-    b3 = rampdouble(0,2,20,200,n1,n2,n3);   assertEqual(b3,mul(2,a3));
+    b3 = rampfloat(0,2,20,200,n1,n2,n3);    assertEqual(b3,mul(2,a3));
                                             assertEqual(b3,mul(a3,2));
 
-    b1 = new double[] {
-      pow(0.0d,2.0d),
-      pow(1.0d,2.0d),
-      pow(2.0d,2.0d),
-      pow(3.0d,2.0d)
+    b1 = new float[] {
+      pow(0.0f,2.0f),
+      pow(1.0f,2.0f),
+      pow(2.0f,2.0f),
+      pow(3.0f,2.0f)
     };
 
-    b2 = new double[n2][n1];     a2 = new double[n2][n1];
-    b3 = new double[n3][n2][n1]; a3 = new double[n3][n2][n1];
+    b2 = new float[n2][n1];     a2 = new float[n2][n1];
+    b3 = new float[n3][n2][n1]; a3 = new float[n3][n2][n1];
 
     for (int i2=0; i2<n2; ++i2) {
       b2[i2] = copy(b1);
@@ -297,9 +306,9 @@ public class ArrayMathTestDouble extends ArrayMathTest {
 
   @Test
   public void testNaturalLog() {
-    b1 = new double[] { Double.NEGATIVE_INFINITY,log(1),log(2),log(3) };
-    b2 = new double[n2][n1];     a2 = new double[n2][n1];
-    b3 = new double[n3][n2][n1]; a3 = new double[n3][n2][n1];
+    b1 = new float[] { Float.NEGATIVE_INFINITY,log(1),log(2),log(3) };
+    b2 = new float[n2][n1];     a2 = new float[n2][n1];
+    b3 = new float[n3][n2][n1]; a3 = new float[n3][n2][n1];
 
     for (int i2=0; i2<n2; ++i2) {
       b2[i2] = copy(b1);
@@ -311,21 +320,16 @@ public class ArrayMathTestDouble extends ArrayMathTest {
       a3[i3] = copy(a2);
     }
 
-    assertAlmostEqual(b1,log(a1));
-    assertAlmostEqual(a1,exp(b1));
-
-    assertAlmostEqual(b2,log(a2));
-    assertAlmostEqual(a2,exp(b2));
-
-    assertAlmostEqual(b3,log(a3));
-    assertAlmostEqual(a3,exp(b3));
+    assertEqual(b1,log(a1));  assertEqual(a1,exp(b1));
+    assertEqual(b2,log(a2));  assertEqual(a2,exp(b2));
+    assertEqual(b3,log(a3));  assertEqual(a3,exp(b3));
   }
 
   @Test
   public void testSquareRoot() {
-    b1 = new double[] { sqrt(0.0d), sqrt(1.0d), sqrt(2.0d), sqrt(3.0d) };
-    b2 = new double[n2][n1];     a2 = new double[n2][n1];
-    b3 = new double[n3][n2][n1]; a3 = new double[n3][n2][n1];
+    b1 = new float[] { sqrt(0.0f), sqrt(1.0f), sqrt(2.0f), sqrt(3.0f) };
+    b2 = new float[n2][n1];     a2 = new float[n2][n1];
+    b3 = new float[n3][n2][n1]; a3 = new float[n3][n2][n1];
 
     for (int i2=0; i2<n2; ++i2) {
       b2[i2] = copy(b1);
@@ -337,22 +341,22 @@ public class ArrayMathTestDouble extends ArrayMathTest {
       a3[i3] = copy(a2);
     }
 
-    assertAlmostEqual(b1,sqrt(a1));  assertAlmostEqual(a1,mul(b1,b1));
-    assertAlmostEqual(b2,sqrt(a2));  assertAlmostEqual(a2,mul(b2,b2));
-    assertAlmostEqual(b3,sqrt(a3));  assertAlmostEqual(a3,mul(b3,b3));
+    assertEqual(b1,sqrt(a1));  assertAlmostEqual(a1,mul(b1,b1));
+    assertEqual(b2,sqrt(a2));  assertAlmostEqual(a2,mul(b2,b2));
+    assertEqual(b3,sqrt(a3));  assertAlmostEqual(a3,mul(b3,b3));
   }
 
   @Test
   public void testPow() {
-    b1 = new double[] {
-      pow(0.0d,2.0d),
-      pow(1.0d,2.0d),
-      pow(2.0d,2.0d),
-      pow(3.0d,2.0d)
+    b1 = new float[] {
+      pow(0.0f,2.0f),
+      pow(1.0f,2.0f),
+      pow(2.0f,2.0f),
+      pow(3.0f,2.0f)
     };
 
-    b2 = new double[n2][n1];     a2 = new double[n2][n1];
-    b3 = new double[n3][n2][n1]; a3 = new double[n3][n2][n1];
+    b2 = new float[n2][n1];     a2 = new float[n2][n1];
+    b3 = new float[n3][n2][n1]; a3 = new float[n3][n2][n1];
 
     for (int i2=0; i2<n2; ++i2) {
       b2[i2] = copy(b1);
@@ -364,16 +368,27 @@ public class ArrayMathTestDouble extends ArrayMathTest {
       a3[i3] = copy(a2);
     }
 
-    assertAlmostEqual(b1,pow(a1,2.0d)); assertAlmostEqual(a1,sqrt(b1));
-    assertAlmostEqual(b2,pow(a2,2.0d)); assertAlmostEqual(a2,sqrt(b2));
-    assertAlmostEqual(b3,pow(a3,2.0d)); assertAlmostEqual(a3,sqrt(b3));
+    assertEqual(b1,pow(a1,2.0f)); assertAlmostEqual(a1,sqrt(b1));
+    assertEqual(b2,pow(a2,2.0f)); assertAlmostEqual(a2,sqrt(b2));
+    assertEqual(b3,pow(a3,2.0f)); assertAlmostEqual(a3,sqrt(b3));
+  }
+
+  @Test
+  public void testIndexAndValueMaxes() {
+    int[] imax = {-1,-1,-1};
+    float rmax = max(a3,imax);
+
+    assertTrue(rmax == a3[n3-1][n2-1][n1-1]);
+    assertEq(n1-1,imax[0]);
+    assertEq(n2-1,imax[1]);
+    assertEq(n3-1,imax[2]);
   }
 
   @Test
   public void testMinMax() {
     Random r = new Random();
-    double min = -1d;
-    double max = 1000d;
+    float min = -1.0f;
+    float max = 1000.0f;
     int j3 = r.nextInt(n3-1)+1;
     int j2 = r.nextInt(n2-1)+1;
     int j1 = r.nextInt(n1-1)+1;
@@ -388,9 +403,9 @@ public class ArrayMathTestDouble extends ArrayMathTest {
     a2[j2][j1]     = min; a2[j2][k1]     = max;
     a1[j1]         = min; a1[k1]         = max;
 
-    double min3 = min(a3,imin3), max3 = max(a3,imax3);
-    double min2 = min(a2,imin2), max2 = max(a2,imax2);
-    double min1 = min(a1,imin1), max1 = max(a1,imax1);
+    float min3 = min(a3,imin3), max3 = max(a3,imax3);
+    float min2 = min(a2,imin2), max2 = max(a2,imax2);
+    float min1 = min(a1,imin1), max1 = max(a1,imax1);
 
     assertEq(min,min3);    assertEq(min,min2);    assertEq(min,min1);
     assertEq(j1,imin3[0]); assertEq(j1,imin2[0]); assertEq(j1,imin1[0]);
@@ -411,7 +426,7 @@ public class ArrayMathTestDouble extends ArrayMathTest {
     assertTrue(isRegular(a2));
     assertTrue(isRegular(a3));
 
-    a2[0] = new double[3];
+    a2[0] = new float[3];
     a3[0] = a2;
 
     assertFalse(isRegular(a2));
@@ -437,43 +452,44 @@ public class ArrayMathTestDouble extends ArrayMathTest {
     assertFalse(isMonotonic(a1));
   }
 
-  private static void assertOnlyContains(double val,double[][][] a) {
+  private static void assertOnlyContains(float val,float[][][] a) {
     for (int i3=0; i3<a.length; ++i3)
       assertOnlyContains(val,a[i3]);
   }
 
-  private static void assertOnlyContains(double val,double[][] a) {
+  private static void assertOnlyContains(float val,float[][] a) {
     for (int i2=0; i2<a.length; ++i2)
       assertOnlyContains(val,a[i2]);
   }
 
-  private static void assertOnlyContains(double val,double[] a) {
+  private static void assertOnlyContains(float val,float[] a) {
     for (int i1=0; i1<a.length; ++i1)
       assertEquals(val,a[i1]);
   }
 
-  private static void assertArraySize(double[][][] a,int n1,int n2,int n3) {
+  private static void assertArraySize(float[][][] a,int n1,int n2,int n3) {
     assertEquals(n3,a.length);
     assertEquals(n2,a[0].length);
     assertEquals(n1,a[0][0].length);
   }
 
-  private static void assertArraySize(double[][] a,int n1,int n2) {
+  private static void assertArraySize(float[][] a,int n1,int n2) {
     assertEquals(n2,a.length);
     assertEquals(n1,a[0].length);
   }
 
-  private static void assertArraySize(double[] a,int n) {
+  private static void assertArraySize(float[] a,int n) {
     assertEquals(n,a.length);
   }
 
 
-  private double[]     a1;
-  private double[][]   a2;
-  private double[][][] a3;
+  private float[]     a1;
+  private float[][]   a2;
+  private float[][][] a3;
 
-  private double[]     b1;
-  private double[][]   b2;
-  private double[][][] b3;
+  private float[]     b1;
+  private float[][]   b2;
+  private float[][][] b3;
+
 
 }
