+++ core/src/test/java/edu/mines/jtk/util/ArrayMathLongTest.java
@@ -22,21 +22,27 @@ import java.util.Random;
 import static edu.mines.jtk.util.ArrayMath.*;
 import static org.testng.Assert.*;
 
+import static edu.mines.jtk.util.ArrayMathTest.*;
+
 /**
- * Tests int array operations in {@link edu.mines.jtk.util.ArrayMath}.
+ * Tests long array operations in {@link edu.mines.jtk.util.ArrayMath}.
  * @author Chris Engelsma
  * @version 2017.05.31
  */
-public class ArrayMathTestInteger extends ArrayMathTest {
+public class ArrayMathLongTest {
+
+  private static final int n3 = 8;
+  private static final int n2 = 6;
+  private static final int n1 = 4;
 
   @BeforeMethod
   public void setUp() {
-    a1 = rampint(0,1,n1);
-    a2 = rampint(0,1,10,n1,n2);
-    a3 = rampint(0,1,10,100,n1,n2,n3);
-    b1 = zeroint(n1);
-    b2 = zeroint(n1,n2);
-    b3 = zeroint(n1,n2,n3);
+    a1 = ramplong(0L,1L,n1);
+    a2 = ramplong(0L,1L,10L,n1,n2);
+    a3 = ramplong(0L,1L,10L,100L,n1,n2,n3);
+    b1 = zerolong(n1);
+    b2 = zerolong(n1,n2);
+    b3 = zerolong(n1,n2,n3);
   }
 
   @Test
@@ -44,32 +50,33 @@ public class ArrayMathTestInteger extends ArrayMathTest {
     zero(a3);
     assertTrue(equal(a3,b3));
 
-    fill(1,a3);
+    fill(1L,a3);
     assertFalse(equal(a3,b3));
   }
 
   @Test
-  public void testRampIntegers() {
+  public void testRampLongs() {
     assertArraySize(a1,n1);
     assertArraySize(a2,n1,n2);
     assertArraySize(a3,n1,n2,n3);
 
     for (int i1=0; i1<n1; ++i1) {
-      assertEquals(i1,a3[0][0][i1]);
-      assertEquals(i1,a2[0][i1]);
-      assertEquals(i1,a1[i1]);
+      long expected = (long)i1;
+      assertEquals(expected,a3[0][0][i1]);
+      assertEquals(expected,a2[0][i1]);
+      assertEquals(expected,a1[i1]);
     }
   }
 
   @Test
-  public void testFillIntegers() {
-    int val = 1;
+  public void testFillLongs() {
+    long val = 1L;
 
-    assertEqual(zeroint(n1,n2,n3),fillint(0,n1,n2,n3));
+    assertEqual(zerolong(n1,n2,n3),filllong(0L,n1,n2,n3));
 
-    a1 = fillint(val,n1);
-    a2 = fillint(val,n1,n2);
-    a3 = fillint(val,n1,n2,n3);
+    a1 = filllong(val,n1);
+    a2 = filllong(val,n1,n2);
+    a3 = filllong(val,n1,n2,n3);
 
     assertArraySize(a1,n1);
     assertArraySize(a2,n1,n2);
@@ -86,9 +93,9 @@ public class ArrayMathTestInteger extends ArrayMathTest {
     zero(a2);
     zero(a3);
 
-    assertOnlyContains(0,a1);
-    assertOnlyContains(0,a2);
-    assertOnlyContains(0,a3);
+    assertOnlyContains(0L,a1);
+    assertOnlyContains(0L,a2);
+    assertOnlyContains(0L,a3);
   }
 
   @Test
@@ -99,9 +106,9 @@ public class ArrayMathTestInteger extends ArrayMathTest {
     assertArraySize(a2,n1,n2);
     assertArraySize(a3,n1,n2,n3);
 
-    b1 = randint(n1);
-    b2 = randint(n1,n2);
-    b3 = randint(n1,n2,n3);
+    b1 = randlong(n1);
+    b2 = randlong(n1,n2);
+    b3 = randlong(n1,n2,n3);
 
     assertArraySize(b1,n1);
     assertArraySize(b2,n1,n2);
@@ -128,50 +135,50 @@ public class ArrayMathTestInteger extends ArrayMathTest {
     assertEqual(b2,a2);
     assertEqual(b3,a3);
 
-    b1 = copy(n1 - 1,a1);
-    b2 = copy(n1 - 1,n2 - 1,a2);
-    b3 = copy(n1 - 1,n2 - 1,n3 - 1,a3);
-    assertEqual(b1,rampint(0,1,n1 - 1));
-    assertEqual(b2,rampint(0,1,10,n1 - 1,n2 - 1));
-    assertEqual(b3,rampint(0,1,10,100,n1 - 1,n2 - 1,n3 - 1));
+    b1 = copy(n1-1,a1);
+    b2 = copy(n1-1,n2-1,a2);
+    b3 = copy(n1-1,n2-1,n3-1,a3);
+    assertEqual(b1,ramplong(0L,1L,n1-1));
+    assertEqual(b2,ramplong(0L,1L,10L,n1-1,n2-1));
+    assertEqual(b3,ramplong(0L,1L,10L,100L,n1-1,n2-1,n3-1));
 
-    copy(n1 - 1,a1,b1);
-    copy(n1 - 1,n2 - 1,a2,b2);
-    copy(n1 - 1,n2 - 1,n3 - 1,a3,b3);
-    assertEqual(b1,rampint(0,1,n1 - 1));
-    assertEqual(b2,rampint(0,1,10,n1 - 1,n2 - 1));
-    assertEqual(b3,rampint(0,1,10,100,n1 - 1,n2 - 1,n3 - 1));
+    copy(n1-1,a1,b1);
+    copy(n1-1,n2-1,a2,b2);
+    copy(n1-1,n2-1,n3-1,a3,b3);
+    assertEqual(b1,ramplong(0L,1L,n1-1));
+    assertEqual(b2,ramplong(0L,1L,10L,n1-1,n2-1));
+    assertEqual(b3,ramplong(0L,1L,10L,100L,n1-1,n2-1,n3-1));
 
-    b1 = copy(n1 - 1,1,a1);
-    b2 = copy(n1 - 2,n2 - 1,2,1,a2);
-    b3 = copy(n1 - 3,n2 - 2,n3 - 1,3,2,1,a3);
-    assertEqual(b1,rampint(1,1,n1 - 1));
-    assertEqual(b2,rampint(12,1,10,n1 - 1,n2 - 1));
-    assertEqual(b3,rampint(123,1,10,100,n1 - 1,n2 - 1,n3 - 1));
+    b1 = copy(n1-1,1,a1);
+    b2 = copy(n1-2,n2-1,2,1,a2);
+    b3 = copy(n1-3,n2-2,n3-1,3,2,1,a3);
+    assertEqual(b1,ramplong(1L,1L,n1-1));
+    assertEqual(b2,ramplong(12L,1L,10L,n1-1,n2-1));
+    assertEqual(b3,ramplong(123L,1L,10L,100L,n1-1,n2-1,n3-1));
 
-    copy(n1 - 1,1,a1,0,b1);
-    copy(n1 - 2,n2 - 1,2,1,a2,0,0,b2);
-    copy(n1 - 3,n2 - 2,n3 - 1,3,2,1,a3,0,0,0,b3);
-    assertEqual(b1,rampint(1,1,n1 - 1));
-    assertEqual(b2,rampint(12,1,10,n1 - 1,n2 - 1));
-    assertEqual(b3,rampint(123,1,10,100,n1 - 1,n2 - 1,n3 - 1));
+    copy(n1-1,1,a1,0,b1);
+    copy(n1-2,n2-1,2,1,a2,0,0,b2);
+    copy(n1-3,n2-2,n3-1,3,2,1,a3,0,0,0,b3);
+    assertEqual(b1,ramplong(1L,1L,n1-1));
+    assertEqual(b2,ramplong(12L,1L,10L,n1-1,n2-1));
+    assertEqual(b3,ramplong(123L,1L,10L,100L,n1-1,n2-1,n3-1));
 
-    b1 = copy(n1 / 2,0,2,a1);
-    b2 = copy(n1 / 2,n2 / 2,0,0,2,2,a2);
-    b3 = copy(n1 / 2,n2 / 2,n3 / 2,0,0,0,2,2,2,a3);
-    assertEqual(b1,rampint(0,2,n1 / 2));
-    assertEqual(b2,rampint(0,2,20,n1 / 2,n2 / 2));
-    assertEqual(b3,rampint(0,2,20,200,n1 / 2,n2 / 2,n3 / 2));
+    b1 = copy(n1/2,0,2,a1);
+    b2 = copy(n1/2,n2/2,0,0,2,2,a2);
+    b3 = copy(n1/2,n2/2,n3/2,0,0,0,2,2,2,a3);
+    assertEqual(b1,ramplong(0L,2L,n1/2));
+    assertEqual(b2,ramplong(0L,2L,20L,n1/2,n2/2));
+    assertEqual(b3,ramplong(0L,2L,20L,200L,n1/2,n2/2,n3/2));
 
     b1 = copy(a1);
     b2 = copy(a2);
     b3 = copy(a3);
-    copy(n1 - 1,1,a1,1,b1);
-    copy(n1 - 2,n2 - 1,2,1,a2,2,1,b2);
-    copy(n1 - 3,n2 - 2,n3 - 1,3,2,1,a3,3,2,1,b3);
-    assertEqual(b1,rampint(0,1,n1));
-    assertEqual(b2,rampint(0,1,10,n1,n2));
-    assertEqual(b3,rampint(0,1,10,100,n1,n2,n3));
+    copy(n1-1,1,a1,1,b1);
+    copy(n1-2,n2-1,2,1,a2,2,1,b2);
+    copy(n1-3,n2-2,n3-1,3,2,1,a3,3,2,1,b3);
+    assertEqual(b1,ramplong(0L,1L,n1));
+    assertEqual(b2,ramplong(0L,1L,10L,n1,n2));
+    assertEqual(b3,ramplong(0L,1L,10L,100L,n1,n2,n3));
   }
 
   @Test
@@ -203,9 +210,9 @@ public class ArrayMathTestInteger extends ArrayMathTest {
     assertFalse(distinct(a2,b2));
     assertFalse(distinct(a1,b1));
 
-    b3 = new int[n3][n2][n1];
-    b2 = new int[n2][n1];
-    b1 = new int[n1];
+    b3 = new long[n3][n2][n1];
+    b2 = new long[n2][n1];
+    b1 = new long[n1];
     assertTrue(distinct(a3,b3));
     assertTrue(distinct(a2,b2));
     assertTrue(distinct(a1,b1));
@@ -219,8 +226,8 @@ public class ArrayMathTestInteger extends ArrayMathTest {
   @Test
   public void testMinMax() {
     Random r = new Random();
-    int min = -1;
-    int max = 1000;
+    long min = -1L;
+    long max = 1000L;
     int j3 = r.nextInt(n3-1)+1;
     int j2 = r.nextInt(n2-1)+1;
     int j1 = r.nextInt(n1-1)+1;
@@ -235,9 +242,9 @@ public class ArrayMathTestInteger extends ArrayMathTest {
     a2[j2][j1]     = min; a2[j2][k1]     = max;
     a1[j1]         = min; a1[k1]         = max;
 
-    int min3 = min(a3,imin3), max3 = max(a3,imax3);
-    int min2 = min(a2,imin2), max2 = max(a2,imax2);
-    int min1 = min(a1,imin1), max1 = max(a1,imax1);
+    long min3 = min(a3,imin3), max3 = max(a3,imax3);
+    long min2 = min(a2,imin2), max2 = max(a2,imax2);
+    long min1 = min(a1,imin1), max1 = max(a1,imax1);
 
     assertEq(min,min3);    assertEq(min,min2);    assertEq(min,min1);
     assertEq(j1,imin3[0]); assertEq(j1,imin2[0]); assertEq(j1,imin1[0]);
@@ -252,19 +259,6 @@ public class ArrayMathTestInteger extends ArrayMathTest {
     assertEq(max,max(a3)); assertEq(max,max(a2)); assertEq(max,max(a1));
     assertEq(min,min(a3)); assertEq(min,min(a2)); assertEq(min,min(a1));
   }
-
-  @Test
-  public void testRegularity() {
-    assertTrue(isRegular(a2));
-    assertTrue(isRegular(a3));
-
-    a2[0] = new int[3];
-    a3[0] = a2;
-
-    assertFalse(isRegular(a2));
-    assertFalse(isRegular(a3));
-  }
-
   @Test
   public void testIncreasingDecreasingMonotonic() {
     assertTrue(isIncreasing(a1));
@@ -284,43 +278,43 @@ public class ArrayMathTestInteger extends ArrayMathTest {
     assertFalse(isMonotonic(a1));
   }
 
-  private static void assertOnlyContains(int val,int[][][] a) {
+  private static void assertOnlyContains(long val,long[][][] a) {
     for (int i3=0; i3<a.length; ++i3)
       assertOnlyContains(val,a[i3]);
   }
 
-  private static void assertOnlyContains(int val,int[][] a) {
+  private static void assertOnlyContains(long val,long[][] a) {
     for (int i2=0; i2<a.length; ++i2)
       assertOnlyContains(val,a[i2]);
   }
 
-  private static void assertOnlyContains(int val,int[] a) {
+  private static void assertOnlyContains(long val,long[] a) {
     for (int i1=0; i1<a.length; ++i1)
       assertEquals(val,a[i1]);
   }
 
-  private static void assertArraySize(int[][][] a,int n1,int n2,int n3) {
+  private static void assertArraySize(long[][][] a,int n1,int n2,int n3) {
     assertEquals(n3,a.length);
     assertEquals(n2,a[0].length);
     assertEquals(n1,a[0][0].length);
   }
 
-  private static void assertArraySize(int[][] a,int n1,int n2) {
+  private static void assertArraySize(long[][] a,int n1,int n2) {
     assertEquals(n2,a.length);
     assertEquals(n1,a[0].length);
   }
 
-  private static void assertArraySize(int[] a,int n) {
+  private static void assertArraySize(long[] a,int n) {
     assertEquals(n,a.length);
   }
 
 
-  private int[]     a1;
-  private int[][]   a2;
-  private int[][][] a3;
+  private long[]     a1;
+  private long[][]   a2;
+  private long[][][] a3;
 
-  private int[]     b1;
-  private int[][]   b2;
-  private int[][][] b3;
+  private long[]     b1;
+  private long[][]   b2;
+  private long[][][] b3;
 
 }
